BA-2: Adding in more tests.

+++ src/main/java/com/craven/bank_account/audit/AuditService.java
@@ -5,6 +5,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -13,7 +14,7 @@ public class AuditService {
     private final Logger logger = LoggerFactory.getLogger(getClass());
     private int batchNumber = 0;
 
-    public void publishBatch(List<Transaction> currentBatch, Double currentBatchValue) {
+    public void publishBatch(List<Transaction> currentBatch, BigDecimal currentBatchValue) {
         if (!currentBatch.isEmpty()) {
             logger.info("Batch Number {}: ", batchNumber);
             logger.info("Total value: £{}", currentBatchValue);
